Support fp32 training.

+++ Language-Understanding-RoBERTa/bert_adapter/fairseq/adapter_utils.py
@@ -51,7 +51,7 @@ class AdapterLinear(nn.Module):
         self.batch_dim = batch_dim
 
         tensor = torch.ones(())
-        self.weight = nn.Parameter(tensor.new_empty(size=(outdim, indim), dtype=torch.half))
+        self.weight = nn.Parameter(tensor.new_empty(size=(outdim, indim)))
 
         self.register_forward_hook(linear_forward_hook)
         self.register_backward_hook(linear_backward_hook)
